Add ability to reserve bulk space in device vector
A very common pattern in massively parallel computing is to reserve
space for an entire thread block at once and to insert those values
later. This commit adds a `bulk_append_implicit` and
`bulk_append_implicit_unsafe` method to `device_vector`, which allows
clients to do this.

The semantics of `bulk_append_implicit` are fully compliant with the C++
standard starting from C++20; C++ standards below that will work just
fine.  Any program which calls this function and compiles should be well
formed, as defined in P0593R6 and P2590R2.

The `bulk_append_implicit_unsafe` method does the same, but
_technically_ leaves the program in an ill-defined state. Note that this
form of ill-defined memory has been used for donkeys years and is
actually safe. It's just technically not standards compliant.

Also adds `bulk_append`, which adds elements that are either
default-constructed or copies of a given reference object.