Separate SYCL compilation details from vecmem core
The SYCL build system for vecmem suffers from a pretty nasty problem
where it has to be compiled using a SYCL-capable compiler. This means
that an installation of SYCL cannot be conveniently classified by the
single compiler used to build it. Fixing this will take a few more PR's
but a more pressing issue is that the vecmem core library is actually
infected with compilation details from the SYCL library. More
specifically, the core library carries information about the function
used to print in SYCL code, leading to the fact that the core library
can change if the SYCL library is enabled, disabled, or is compiled with
a different compiler which is very undesirable. This commit migrates all
the SYCL-related logic from the core library to the SYCL library to
resolve this issue.