Mark copy events as `nodiscard`
Currently, there is no compile-time mechanism to ensure that copy
objects are properly handled, i.e. either waited for or ignored. This
commit adds the `nodiscard` attribute to methods of the copy objects
which return events to ensure that the compiler will issue a warning if
the object is ignored.