Atomic Reference Reorganization, main branch (2024.08.08.) (#291)
* Split vecmem::device_atomic_ref into multiple classes.

Each new class is meant to be used on just one "platform",
and vecmem::device_atomic_ref is a typedef to the one that
is appropriate in any given situation.

The selection procedure is not correct yet, the "POSIX" version
of the code never gets selected at the moment.

* Introduced vecmem::hip::device_atomic_reference.

It is the same as vecmem::cuda::device_atomic_reference, with an additional
include needed by HIP.

* Added simple test for using the posix atomic reference.

Not in a way yet which would allow the re-evaluation of the
C++ compiler in client projects.

* Tweaked the flags of the atomic_ref functions.

The SYCL implementation never needs __host__ or __device__.

The "POSIX" implementation will only ever work in host code.

The CUDA and HIP implementations are now included more generally
using the __CUDACC__ and __HIPCC__ macros. At which point the
vecmem::cuda::device_atomic_ref implementation had to be tweaked
to only use __threadfence() when building device code.

* Moved vecmem::device_address_space out of details/

Since the main declaration of vecmem::device_atomic_ref
needs that enumeration as well.