Introduced vecmem_check_sycl_source_compiles(...).
It provides an interface that is in line with the
check_cxx_source_compiles(...) function, taking the
code to be tested, from the CMake code directly.
Instead of having to rely on a physical file in the
source tree.

Kept vecmem_check_sycl_code_compiles(...) untouched,
as for one particular test (about whether a polyfill
should be provided for assertions) it is more
appropriate than the new function.

Removed all of the test .sycl files that are no longer
necessary with the new function.