SoA Proxy, main branch (2024.09.25.) (#296)
* Introduced a concatenation constructor for vecmem::tuple.

Did not try to make a fully functional vecmem::tuple_cat
function, as that was not technically needed so far.

* Introduced vecmem::edm::proxy for host containers.

Proxies can be used to provide an "AoS view" over an
SoA container. A view that would make use of the same
interface that the container itself also uses.

Code for device container support was also added, but
is probably not functional at the moment.

* Added "proxy functionality" to vecmem::edm::device as well.

While also ensuring that vecmem::edm::proxy would be usable in
device code, and also with C++14.

* Tweaked the new code and its documentation a little.

* Switched to references with the const proxies.

As feared, this premature/misguided optimization was not helping
with code performance in real life.