Make several improvements to binary page allocator
This commit makes two main changes to improve the performance of the
binary page memory resource.

Firstly, it makes the allocator more eager to unsplit (i.e. merge) pages
so that they can be reused for new allocation, which should reduce the
memory footprint of when using very large allocations.

Secondly, it sets a minimum size of allocations per superpage, which
depends on the size of the superpage. This prevents very small
allocations from ending up in massive blocks (increasing fragmentation)
and also decreases the total number of pages (increasing performance).