Make vecmem::tuple trivially copyable.
As it turns out, this is a requirement from Alpaka. Even if none
of CUDA, HIP or SYCL require this about kernel parameters, Alpaka
does require this about all the parameters that it receives.

This lead me down a rabbit hole, resulting in a number of updates
in how vecmem::tuple would work. I believe making it a lot more
correct in the process.

Cleared up all reported SonarCloud issues in the tuple code
as a sideproduct.