SoA copy::to, main branch (2025.05.30.) (#334)
* Introduced vecmem::data::get_capacities(...).

It is a convenient way for accessing the capacities of a jagged vector.

Fixed the constructor of jagged_vector_buffer, to make use of this
new function instead of the one it had so far. (Which would only
work with non-resizable and/or host-allocated buffers.)

* Introduced vecmem::edm::get_capacities(...).

Leaning on vecmem::data::get_capacities(...), it provides a
convenient way of getting the capacities of the inner vectors of an
SoA container.

* Introduced copy::to(...) for SoA containers.

All while making the other implementations of copy::to(...)
follow the style introduced for the SoA containers.