Empty SoA Buffer Fix, main branch (2025.09.19.) (#341)
* Introduced tests for creating empty SoA buffers.

Triggered by the knowledge that something is going wrong with
zero sized resizable "simple" buffers at the moment.

* Let the size of zero capacitied buffers be cleared.

For buffers that have only 1D vectors, when we create an
instance with zero capacity, we still need to explicitly
set the size of the buffer to zero as part of initialization.