Replaced vecmem::async_value with vecmem::async_size(s).
While at first it seemed like a good idea to use a single type for
handling single and vector sizes, it resulted in the interface being
very inconvenient for single-size retrievals. Since these are the
majority of the retrievals that we do, that was not fortunate.

Now I rather added a separate async_size and async_sizes type.
Both of which would provide a (hopefully) more easy to use interface
for the users.